WEB-3650: Removed races table
