+++ prisma/schema/municipality.jsonTypes.ts
@@ -1,0 +1,15 @@
+export {}
+
+declare global {
+  export namespace PrismaJson {
+    export type MunicipalityData = {
+      population?: string
+      density?: string
+      income_household_median?: string
+      unemployment_rate?: string
+      home_value?: string
+      county_name?: string
+      city?: string
+    }
+  }
+}
