--- prisma/schema/race.prisma
@@ -1,29 +1,0 @@
-enum LevelType {
-  city
-  county
-  state
-  federal
-}
-
-model Race {
-  id             Int           @id @default(autoincrement())
-  createdAt      DateTime      @default(now()) @map("created_at")
-  updatedAt      DateTime      @updatedAt @map("updated_at")
-  ballotId       String        @unique @map("ballot_id")
-  ballotHashId   String?       @map("ballot_hash_id")
-  hashId         String        @unique @map("hash_id")
-  positionSlug   String        @map("position_slug")
-  state          String        @db.Char(2)
-  electionDate   DateTime?     @map("election_date")
-  level          LevelType
-  subAreaName    String        @map("sub_area_name")
-  subAreaValue   String        @map("sub_area_value")
-  /// [RaceData]
-  data           Json          @default("{}") @db.JsonB
-  county         County?       @relation(fields: [countyId], references: [id])
-  countyId       Int?          @map("county_id")
-  municipality   Municipality? @relation(fields: [municipalityId], references: [id])
-  municipalityId Int?          @map("municipality_id")
-
-  @@map("race")
-}
