+++ src/elections/types/elections.types.ts
@@ -21,3 +21,24 @@ export type PrimaryElectionDates = {
     primaryElectionId: string
   }
 }
+
+export type GeoData = {
+  name: string
+  type: string
+  city?: string
+  county?: string
+  state?: string
+  township?: string
+  town?: string
+  village?: string
+  borough?: string
+}
+
+export enum OfficeLevel {
+  FEDERAL = 10,
+  STATE = 8,
+  COUNTY = 6,
+  CITY = 4,
+  LOCAL = 0,
+  DEFAULT = 12,
+}
